+++ modules/project/core/src/sap/commerce/toolset/project/actionSystem/Util.kt
@@ -35,6 +35,6 @@ fun Project.triggerAction(
                 it, dataContext.invoke(),
                 null, place, uiKind, null
             );
-            ActionUtil.performActionDumbAwareWithCallbacks(it, event)
+            ActionUtil.performAction(it, event)
         }
 }