+++ modules/project/import-ui/src/sap/commerce/toolset/project/actionSystem/ProjectRefreshAction.java
@@ -70,7 +70,7 @@ public class ProjectRefreshAction extends AnAction {
                 null
             );
 
-            ActionUtil.performActionDumbAwareWithCallbacks(action, actionEvent);
+            ActionUtil.performAction(action, actionEvent);
         }, ModalityState.nonModal());
     }