+++ modules/typeSystem/diagram/src/sap/commerce/toolset/typeSystem/actionSystem/ExcludeTypeNameDiagramAction.kt
@@ -46,7 +46,7 @@ class ExcludeTypeNameDiagramAction : DiagramAction(
             }
 
             val action = ActionManager.getInstance().getAction("Diagram.RefreshDataModelManually")
-            ActionUtil.performActionDumbAwareWithCallbacks(action, event)
+            ActionUtil.performAction(action, event)
         }
     }