+++ modules/typeSystem/diagram/src/sap/commerce/toolset/typeSystem/actionSystem/ExpandAllDiagramAction.kt
@@ -40,7 +40,7 @@ class ExpandAllDiagramAction : DiagramAction(
                 it.expandAllNodes()
 
                 val action = ActionManager.getInstance().getAction("Diagram.RefreshDataModelManually")
-                ActionUtil.performActionDumbAwareWithCallbacks(action, event)
+                ActionUtil.performAction(action, event)
             }
     }